+++ R/run_validation.R
@@ -249,6 +249,19 @@ run_input_validation <- function(path = getwd(),
            }
   )
 
+  tryCatch(check_CUI_license_match(path = path,
+                                 filename = filename,
+                                 sheet_name = sheet),
+           error = function(e) {
+             err_count <<- err_count +1
+             cli::cli_bullets(c(e$message, e$body))
+           },
+           warning = function(w) {
+             warn_count <<- warn_count + 1
+             cli::cli_bullets(c(w$message, w$body))
+           }
+  )
+
   tryCatch(check_prod_units(path = path,
                             filename = filename,
                             sheet_name = sheet),